WDIC: refactring wdic node handling
